pyproject.toml
[project]
name = "agent-forge"
version = "0.6.0"
description = "Multi-agent orchestration framework — assemble AI teams that collaborate, debate, and create"
requires-python = ">=3.10"
dependencies = [
    "rich>=13.0.0",
    "edge-tts>=7.0.0",
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
# If you only want one provider, install just the core + one of these.
anthropic = ["anthropic>=0.40.0"]
google = ["google-genai>=1.0.0"]
